requirements.txt
python-decouple==3.7
Telethon==1.25.0
requests==2.28.1
